refactor : User, Search 컨트롤러 분리
