+++ src/main/java/com/woowacourse/zzinbros/user/domain/repository/UserRepository.java
@@ -6,9 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-
     boolean existsUserByEmailEmail(String email);
 
     Optional<User> findByEmailEmail(String email);
-
 }
