+++ src/main/java/com/woowacourse/zzinbros/user/web/controller/SearchController.java
@@ -1,0 +1,28 @@
+package com.woowacourse.zzinbros.user.web.controller;
+
+import com.woowacourse.zzinbros.user.dto.UserResponseDto;
+import com.woowacourse.zzinbros.user.service.SearchService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import java.util.Set;
+
+@Controller
+@RequestMapping("/users")
+public class SearchController {
+    private final SearchService searchService;
+
+    public SearchController(SearchService searchService) {
+        this.searchService = searchService;
+    }
+
+    @GetMapping
+    public ResponseEntity<Set<UserResponseDto>> search(@RequestParam("name") String name) {
+        Set<UserResponseDto> users = searchService.search(name);
+        return new ResponseEntity<>(users, HttpStatus.OK);
+    }
+}
